[ty] Move uv command execution into its own module
